Added support for orthographic projection for RenderTriangles and TestRect, both use normalized depth. Only StaticLib compiles, because that's all I need.
For ortho, very rarely, individual triangle seems to get culled instead of clipped, don't know if this is an old bug now visible, or a new one. Seems to be working fine overall.

Disabled QUICK_MASK, because it produces a lot of holes.

Disabled LTO, because with GD3D11 (that itself has LTO) there is no performance difference, and link times are ~15x slower. MSVC is unable to constant fold through virtual calls, so that's probably why.
Changed target name of StaticLib to "MaskedOcclusionCulling".
Enabled C++17, because I used "if constexpr".